% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads_at_groups.R
\name{count_reads_at_groups}
\alias{count_reads_at_groups}
\title{A function that counts the reads at specific positions in each sample}
\usage{
count_reads_at_groups(bamList, gRangesToLookFor)
}
\arguments{
\item{bamList}{A list of bams produced by parallel_read_bams}

\item{gRangesToLookFor}{a GRanges object. Reads corresponding to the ranges
in this object will be found from each of the bam files}
}
\value{
Returns a data frame that shows the number of reads from each file
mapping within the ranges specified in gRangesToLookFor
}
\description{
This function is intended to use the GRanges output from
find_related_positions to count the number of reads that occur in groups of
positions that looked interesting across the samples but can use any GRanges
object you're interested in, for example, a start codon only GFF. This
function defaults to 4 threads.
}
