% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_related_positions.R
\name{find_related_positions}
\alias{find_related_positions}
\title{Finds genomic positions that have many reads mapping to them.}
\usage{
find_related_positions(bamDf, maxDiff, gffLoc)
}
\arguments{
\item{bamDf}{A data frame produced by the tally_reads function}

\item{maxDiff}{an integer, the maximum distance in number of bases to
consider something as "close". If a read occurs at position 1 and another at
position 4, these reads will be in the same "group" if maxDiff >= 3 but in
separate groups if maxDist <= 2}

\item{gffLoc}{the location of a gff}
}
\value{
a GRanges object that shows positions that have many reads at them
across all samples.
}
\description{
This function helps the user find positions that have many reads mapping to
them across all their samples under the logic that positions with many reads
mapping to them in at least one sample (though not necessarily others) could
be important.
}
