% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_read_bams.R
\name{parallel_read_bams}
\alias{parallel_read_bams}
\title{Read in BAM files in a parallel manner.}
\usage{
parallel_read_bams(dirWithBams, ncores)
}
\arguments{
\item{dirWithBams}{path to a directory with bam files in it, it will only.
attempt to read things that end in .bam.}

\item{ncores}{the number of cores to use, typically 1 per bam.}
}
\value{
A list of bams read in using readGAlignments.
}
\description{
This function simply reads BAM files in using parallel's mclapply and
GenomicAlignments readGAlignments. It results in a named list of bam files.
By default, this reads all BAMs it can find, if you want to use a subset of
BAMs from a folder, you make your own named list of BAM files.
}
