% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_contam_reads.R
\name{remove_contam_reads}
\alias{remove_contam_reads}
\title{Remove contaminating reads in a parallel fashion.}
\usage{
remove_contam_reads(bamList, gffLoc, contaminants)
}
\arguments{
\item{bamList}{A named list of bam files}

\item{gffLoc}{The path to a gff}

\item{contaminants}{A character vector containing the thing in gffCol you
want to remove, for example, rRNA. This is searched for in the gff "type"
column to make a gff that defines the contaminating reads. This is currently
not flexible because it requires a "type" column.}
}
\value{
A list of bams depleted of contaminating reads
}
\description{
This function removes contaminating reads, generally rRNA/tRNA reads, but
can be configured to remove reads mapping to wherever the user chooses.
It uses a user provided GFF to remove reads that map to user designated
regions.
}
