% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_reads.R
\name{tally_reads}
\alias{tally_reads}
\title{Count reads at positions given an offset}
\usage{
tally_reads(bamList, offset, ncores)
}
\arguments{
\item{bamList}{A named list of bam files, presumably from the remove_contam_reads function}

\item{offset}{A data frame of columns qwidth, offset, specifying the read length specific offsets}
}
\value{
A single data frame with counts per position per read length for all samples
}
\description{
This function counts the number of reads with 3' ends at a given position and
separates them by read lengths.
}
