% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_reads.R
\name{tally_reads}
\alias{tally_reads}
\title{Count reads at positions given an offset}
\usage{
tally_reads(bamList, offset)
}
\arguments{
\item{bamList}{A named list of bam files, presumably from the remove_contam_reads function}

\item{offset}{The offset from the 3' end, this is applied to all reads of all lengths}
}
\value{
A single data frame with counts per position per read length for all samples
}
\description{
This function counts the number of reads with 3' ends at a given position and
separates them by read lengths. This function defaults to using 4 threads. IT
CURRENTLY ONLY CONSIDERS POSITIVE STRAND READS.
}
